Making Player an interface.
